warehouse servisleri olusturuldu.

+++ maindata-service/src/main/java/com/oee/entity/ClientInfo.java
@@ -24,7 +24,7 @@ public class ClientInfo {
 	@Id
 	@Column(name="CLIENT_ID")
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer clientId;
+	private Long clientId;
 	
 	@Column(name="CLIENT_NAME")
 	private String clientName;
@@ -68,13 +68,13 @@ public class ClientInfo {
 		this.plants = plants;
 	}
 
-	public Integer getClientId() {
+	public Long getClientId() {
 		return clientId;
 	}
 
-	public void setClientId(Integer clientId) {
+	public void setClientId(Long clientId) {
 		this.clientId = clientId;
 	}
-	
+
 	
 }
